feat: info tooltip for *fixed* option in recurrence transaction
